Strip hyphenated words and word-parts
When checking a hyphenated word, the whole word is checked against the
stoplist and then individual parts are checked.

The passed through text might be a bit strange if only part of a word is
stripped, but this is really unavoidable.  For example, with the
stopword "foo", the text C<'foo-of-a-kind'> will get stripped to
C<'of-a-kind'>.  For the purposes of spell checking, this is probably
fine.

Fixes #9.