revise/improve word detection again
The new approach strips leading punctuation, then keeps everything up to
the first non-period, non-single-quote, non-colon trailing punctuation mark.

At that point we have mixed characters, periods, single-quotes and
colons.  We strip all trailing quotes, periods and colons, and are left
with a word that could be a possessive ("John's"), a contraction
("don't"), an abbreviation lacking the trailing period ("Ph.D") or
some sort of possibly code-like thing.

We then strip possessives.

We then drop any word that starts with a sigil or has internal symbols
as likely to be code without C<> markup protecting it.

Anything remaining must have at least one character (\w) in it or the
word is dropped as garbage somehow left over from the other munging.