fix handling of spaces in C<> and F<> tags
The C<> and F<> tags are meant to be removed from the final output, but
if they are directly next to other text, are meant to add an underscore
to that text, causing them to be parsed as a perl token and ignored.

The regexes for spaces were operating on $_, an internal value in
Pod::Parser, rather than $seq_arg, the actual content of the tag. The
value of $_ would be almost, but not exactly the same. It would include
the ending > from the tag, and if there was an internal tag (like E<>)
it would omit that tag and anything before it. Including the end tag
would mean that it would never match the ending whitespace, and skipping
past interior tags would sometimes accidentally find starting whitespace
that shouldn't have matched.

Correct the regexes to operate on the correct variable, and add a test
for the result.