refactor wordlist loading
Rather than calculating the wordlist path in a constant, then reading it
later, just do it all at compile time. Since we aren't saving the path
for later, we don't care if it's an absolute path, and don't need to
Cwd::abs_path it. The Cwd::abs_path call had to be moved anyway, since
it only works on files that exist.